requirements.txt
anyio==3.3.2
asgiref==3.4.1
click==8.0.1
colorama==0.4.4
fastapi==0.70.0
h11==0.12.0
httptools==0.2.0
idna==3.2
pydantic==1.8.2
PyJWT==2.2.0
PyMySQL==1.0.2
python-dotenv==0.19.0
PyYAML==5.4.1
sniffio==1.2.0
starlette==0.16.0
typing-extensions==3.10.0.2
uvicorn==0.15.0
watchgod==0.7
websockets==10.0
